Removed RascalFileSystemInVSCode on the java side
